+++ lib/app_other/app_string.dart
